% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-kvi.R
\name{.kvi}
\alias{.kvi}
\title{.kvi}
\usage{
.kvi(x, k = TRUE, v = TRUE, i = TRUE)
}
\arguments{
\item{x}{object such as vector or list to separate into key value pairs}

\item{k}{logical; whether to output (k)ey}

\item{v}{logical; whether to output (v)alue}

\item{i}{logical; whether to output (i)ndex}
}
\description{
Internal function for producing list-of-lists
}
\details{
\strong{DO NOT DIRECTLY USE THIS FUNCTION}. Use \code{\link[=kvi]{kvi()}}, \code{\link[=kv]{kv()}}, instead.

Produces list-of-list where the internal lists contain keys, values or
indices. Useful for iterating. See \code{\link[=kv]{kv()}}.
}
