% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/each_row.R
\name{each_row}
\alias{each_row}
\title{Iterate over each_row of a table}
\usage{
each_row(table)
}
\arguments{
\item{table}{to iterate over the rows}
}
\description{
Iterate over each row
}
\details{
creates a list object for each row of a table
}
\examples{

each_row( iris[1:3, ])

for( . in each_row( iris[1:5,] ) )
  print(.$Species)


}
