% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kv.R, R/kvi.R
\name{kv}
\alias{kv}
\alias{kv.default}
\alias{kvi}
\alias{kvi.default}
\alias{ki}
\alias{ki.default}
\alias{vi}
\alias{vi.default}
\title{Key-value-index iteration}
\usage{
kv(x, ...)

\method{kv}{default}(x, ...)

kvi(x, ...)

\method{kvi}{default}(x, ...)

ki(x, ...)

\method{ki}{default}(x, ...)

vi(x, ...)

\method{vi}{default}(x, ...)

kvi(x, ...)

\method{kvi}{default}(x, ...)
}
\arguments{
\item{x}{object such as vector or list to separate into key value pairs}

\item{...}{additional arguments for other S3 functions, currently unused.}
}
\value{
A named list-of-lists; each element of \code{x} becomes a one element list
with elements \code{k} and \code{v} representing the keys and value respectively.
}
\description{
Creates a list with \code{k} and \code{v} elements useful for iteration in loops.
}
\details{
No magic here.

These functions convert each element of \code{x} into a list contain a (k)ey,
(v)alue or (i)ndex.

For \strong{each element of \code{x}}, a list with elements \code{k} and \code{v}:
\code{k} is the key(name) for the element. Missing names are made \code{NULL}
\code{v} is that value of that element.

\code{kv} works for vector and list-like objects including tables.

For many cases, key-value iteration can be done with \code{*apply} or
\code{paste} functions. This is made to be explicit and work on a variery of
objects.

\code{kv} is an S3 generic function.

\code{kvi} adds an index to element index to the resulting list so the
index iteration can also be tracked.
}
\examples{

  mylist = list( a=1, b=2, c=3 )

  for( . in kv(mylist))
    cat( "key is: ", .$k, "| value is:", .$v, "\\n" )

 # Lists
 li <- list(a=1,b=2,c=3)
 kv(li)

 # vectors
 v <- c(a=1, b=2, c=3 )
 kv(li)



x <- c( a="Foo", b="Bar" )
kvi(x)

for( . in kvi(x) )
  cat( .$i, .$k, .$v, "\\n")

}
\references{
\itemize{
\item \href{http://stackoverflow.com/questions/18572921/for-loop-in-r-with-key-value}{SO: for-loop-in-r-with-key-value}
\item \href{http://stackoverflow.com/questions/4500106/iterate-over-key-value-pair-from-a-list}{SO: iterate-over-key-value-pair-from-a-list}
}
}
